init commit!
PicForge is an image processing program. The building is pretty sketchy
atm because I am using batch script. Currently using a standard mvc
pattern to design it. I plan on using jni to implement the image
processing algos (not because speed is super important, but because I
think it's awesome that jni exists).

I genuinely tried to get gradle to build a generic 'hello world' project
that uses jni and could not. I want this to be crossplatform (hence the
desire to use a gradle wrapper) and I also want to use a build system
because using scripts for each independent system sounds kinda hacky.
After some research, it seems that maven is the 'go-to' for java
projects so that's probably what I'll end up using. But for now...
build.bat!