zoomImage() removed
It was a neat little set of functions, but the more it was used, the
more it would blur an image. I decided that images smaller than the
default JFrame size wouldn't resize it. The JFrame will inherit the
images size if the image is larger. this seems like the simplist way
to do things.